requirements.txt
# Core Dependencies
pandas>=1.0.0
pm4py>=2.6.0  # Process Mining for Python
matplotlib==3.8.2

# Web Framework
fastapi
uvicorn[standard]

# Database
psycopg2-binary  # PostgreSQL adapter
pgvector  # Vector database extension

# File Processing
openpyxl  # For reading XLSX files in pandas
python-docx  # For reading DOCX files
PyPDF2  # For PDF processing
chardet  # Character encoding detection
python-multipart
python-dotenv  # For loading environment variables
beautifulsoup4

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0  # Coverage reporting
pytest-asyncio>=0.21.0  # Async test support
httpx>=0.24.0  # HTTP client for testing
requests>=2.28.0  # HTTP library

# AI/ML (for embedding service)
sentence-transformers>=2.2.0  # Text embeddings
numpy>=1.20.0